Ajustes repositório base SQL

+++ ConciliadorFinanceiro.API/Controllers/LancamentoFinanceiroController.cs
@@ -53,7 +53,7 @@ namespace ConciliadorFinanceiro.API.Controllers
 
             var lancamentoAtualizar = await _businessLancamento.Consultar(lancamentoFinanceiro);
 
-            if (lancamentoAtualizar.Status == StatusLancamento.Conciliado)
+            if (lancamentoAtualizar.Status == (int)StatusLancamento.Conciliado)
                 return BadRequest();
 
             lancamentoAtualizar.Valor = lancamentoFinanceiro.Valor;
